Object unification, B-trees instead of hashes
This is what I get for not committing to a project for too long -- a
bunch of random changes that end up in one commit.

I implemented the (rather headache-inducing) object unification
algorithm, so now most JSON can be processed with no problems at all.

I also switched everything over from using hash maps and vectors to
using B-tree maps and sets. This makes a whole lot more sense, because
it means we have deterministic orderings on the output structures by
default.